ak2: introduce 'lkt' binary for tuning the Kernel
We'll use the new spectrum app to run the property that will run the
script sourcing the lambda.preset symlinking to the actual presets from
spectrum proerties.

This creates an "API" for tuning and creates exposed values that can be
changed from userspace at will.

This commit creates the house for that logic.

Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>