ak2: lkt binary (script) driven Spectrum support
Made for convenience to expose all the tuning parameters to a unique
tuning API.

This doesn't depend on spectrum since it uses a persist property and the
values aren't hard-coded any way.

Everytime the property is changed from spectrum app, it will tune the
Kernel accordingly as the *.profile values want.

There's also the possility of chaging the profile with lkt itself,
directly.

Some clean up and logic improvements may be needed.

TODO:
Allow tuning the profiles to create a custom one, "fork" of the presets.

Inspired by @dorimanx work at STweaks profiles.

Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>